+++ grpc-google-cloud-bigquerystorage-v1beta2/src/main/java/com/google/cloud/bigquery/storage/v1beta2/BigQueryWriteGrpc.java
@@ -23,6 +23,9 @@ import static io.grpc.MethodDescriptor.generateFullMethodName;
  * <pre>
  * BigQuery Write API.
  * The Write API can be used to write data to BigQuery.
+ * The [google.cloud.bigquery.storage.v1
+ *   API](/bigquery/docs/reference/storage/rpc/google.cloud.bigquery.storage.v1)
+ *   should be used instead of the v1beta2 API for BigQueryWrite operations.
  * </pre>
  */
 @javax.annotation.Generated(
@@ -33,7 +36,8 @@ public final class BigQueryWriteGrpc {
 
   private BigQueryWriteGrpc() {}
 
-  public static final String SERVICE_NAME = "google.cloud.bigquery.storage.v1beta2.BigQueryWrite";
+  public static final java.lang.String SERVICE_NAME =
+      "google.cloud.bigquery.storage.v1beta2.BigQueryWrite";
 
   // Static method descriptors that strictly reflect the proto.
   private static volatile io.grpc.MethodDescriptor<
@@ -369,6 +373,9 @@ public final class BigQueryWriteGrpc {
    * <pre>
    * BigQuery Write API.
    * The Write API can be used to write data to BigQuery.
+   * The [google.cloud.bigquery.storage.v1
+   *   API](/bigquery/docs/reference/storage/rpc/google.cloud.bigquery.storage.v1)
+   *   should be used instead of the v1beta2 API for BigQueryWrite operations.
    * </pre>
    */
   public interface AsyncService {
@@ -502,6 +509,9 @@ public final class BigQueryWriteGrpc {
    * <pre>
    * BigQuery Write API.
    * The Write API can be used to write data to BigQuery.
+   * The [google.cloud.bigquery.storage.v1
+   *   API](/bigquery/docs/reference/storage/rpc/google.cloud.bigquery.storage.v1)
+   *   should be used instead of the v1beta2 API for BigQueryWrite operations.
    * </pre>
    */
   public abstract static class BigQueryWriteImplBase
@@ -519,6 +529,9 @@ public final class BigQueryWriteGrpc {
    * <pre>
    * BigQuery Write API.
    * The Write API can be used to write data to BigQuery.
+   * The [google.cloud.bigquery.storage.v1
+   *   API](/bigquery/docs/reference/storage/rpc/google.cloud.bigquery.storage.v1)
+   *   should be used instead of the v1beta2 API for BigQueryWrite operations.
    * </pre>
    */
   public static final class BigQueryWriteStub
@@ -670,6 +683,9 @@ public final class BigQueryWriteGrpc {
    * <pre>
    * BigQuery Write API.
    * The Write API can be used to write data to BigQuery.
+   * The [google.cloud.bigquery.storage.v1
+   *   API](/bigquery/docs/reference/storage/rpc/google.cloud.bigquery.storage.v1)
+   *   should be used instead of the v1beta2 API for BigQueryWrite operations.
    * </pre>
    */
   public static final class BigQueryWriteBlockingStub
@@ -773,6 +789,9 @@ public final class BigQueryWriteGrpc {
    * <pre>
    * BigQuery Write API.
    * The Write API can be used to write data to BigQuery.
+   * The [google.cloud.bigquery.storage.v1
+   *   API](/bigquery/docs/reference/storage/rpc/google.cloud.bigquery.storage.v1)
+   *   should be used instead of the v1beta2 API for BigQueryWrite operations.
    * </pre>
    */
   public static final class BigQueryWriteFutureStub
@@ -1026,9 +1045,9 @@ public final class BigQueryWriteGrpc {
   private static final class BigQueryWriteMethodDescriptorSupplier
       extends BigQueryWriteBaseDescriptorSupplier
       implements io.grpc.protobuf.ProtoMethodDescriptorSupplier {
-    private final String methodName;
+    private final java.lang.String methodName;
 
-    BigQueryWriteMethodDescriptorSupplier(String methodName) {
+    BigQueryWriteMethodDescriptorSupplier(java.lang.String methodName) {
       this.methodName = methodName;
     }